// auth/features/ory-elements/types.ts
export type OryNodeButtonButtonProps = {
  type: "button" | "submit" | "reset"
  name: string
  value: string | number | readonly string[] | undefined
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  onClick: (event: any) => void
  disabled?: boolean
}

export type OryNodeButtonProps = {
  /** @deprecated Use node.attributes instead. */
  attributes: UiNodeInputAttributes
  node: UiNodeInput
  /**
   * Indicates whether the form is currently being submitted via this button.
   */
  isSubmitting: boolean

  buttonProps: OryNodeButtonButtonProps
}